+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## [1.2.2] - 2026-22-04
+### Changed
+- Moved minification function in `ui/main_window.py` to `core/worker.py`
+- Minification function now runs on a Thread in `ui/main_window.py` as to not freeze window while active
+
 ## [1.2.1] - 2026-22-04
 ### Added
 - File size overview (original size vs minified size & reduction %)